requirements.txt
azure-cosmos
python-dotenv
